+++ curl.go
@@ -1,6 +1,10 @@
 package gorequest
 
-import "moul.io/http2curl"
+import (
+	"net/http"
+
+	"moul.io/http2curl"
+)
 
 // SetCurlCommand enable the curlcommand mode which display a CURL command line.
 func (s *SuperAgent) SetCurlCommand(enable bool) *SuperAgent {
@@ -11,6 +15,7 @@ func (s *SuperAgent) SetCurlCommand(enable bool) *SuperAgent {
 // AsCurlCommand returns a string representing the runnable `curl' command
 // version of the request.
 func (s *SuperAgent) AsCurlCommand() (string, error) {
+	// FIXME: why here MakeRequest again?
 	req, err := s.MakeRequest()
 	if err != nil {
 		return "", err
@@ -20,4 +25,16 @@ func (s *SuperAgent) AsCurlCommand() (string, error) {
 		return "", err
 	}
 	return cmd.String(), nil
+}
+
+func (s *SuperAgent) logCurlCommand(req *http.Request) {
+	if s.CurlCommand {
+		curl, err := http2curl.GetCurlCommand(req)
+		s.logger.SetPrefix("[curl] ")
+		if err != nil {
+			s.logger.Println("Error:", err)
+		} else {
+			s.logger.Printf("CURL command line: %s", curl)
+		}
+	}
 }
