+++ redirect.go
@@ -19,3 +19,11 @@ func (s *SuperAgent) RedirectPolicy(policy func(req Request, via []Request) erro
 	}
 	return s
 }
+
+// DisableRedirect will disable the redirect of status code 3xx.
+func (s *SuperAgent) DisableRedirect() *SuperAgent {
+	s.Client.CheckRedirect = func(req *http.Request, via []*http.Request) error {
+		return http.ErrUseLastResponse
+	}
+	return s
+}
