+++ trace.go
@@ -1,0 +1,8 @@
+package gorequest
+
+import "net/http/httptrace"
+
+func (s *SuperAgent) HttpTrace(trace *httptrace.ClientTrace) *SuperAgent {
+	s.trace = trace
+	return s
+}
